Rewrite most many code 1. remove a lot of unnecessary `/` and `\\` check. 2. test on macOS. 3. remove old build system add a new one (choose from list style).

+++ MakerClass.py
@@ -1,59 +1,56 @@
-import sys
-import os
 import glob
+import os
 import sublime
 
 
 class Makerfile():
 
-    def __init__(self, settings, src=None, obj=None, build=None, header=None):
+    def __init__(self, settings, pwd, is_single_file):
         self.makefile = ""
 
         self.variables = {}
         self.output_file = "output"
         self.string_template = ""
 
-        if sys.platform == "win32":
+        if sublime.platform() == "windows":
             self.output_file += ".exe"
 
-        self.src = src
-        self.obj = obj
-        self.build = build
-        self.header = header
-        self.sources = self.get_source_files(self.src)
         self.settings = settings
-        self.path_sep = '\\'
-
-        if sys.platform.startswith('linux'):
-            self.path_sep = '/'
+        self.src = settings.get('src_dir', '').strip()
+        self.obj = settings.get('obj_dir', '').strip()
+        self.build = settings.get('build_dir', '').strip()
+        self.header = settings.get('include_dir', '')
+        if is_single_file:
+            self.sources = self.get_source_files(pwd)
+        else:
+            self.sources = self.get_source_files(os.path.join(pwd, self.src))
 
     def make_file(self):
-
+        '''Exported API for making final makefile content'''
         if not self.sources:
-            print("No souces provided")
-            return None
+            return '# No sources provided'
 
-        self.makefile += self.handle_variable(self.settings)
+        self.makefile = self.handle_variable(self.settings)
         self.make_string_template()
 
-        self.makefile += "\n"
+        return '\n'.join([
+            self.makefile,
+            self.str_main_file(),
+            '\n' + self.str_recipe_target(),
+            self.str_make_run(),
+            self.str_make_clean()
+        ])
 
-        self.makefile += self.str_main_file()
-
-        self.makefile += "\n\n{}\n".format(self.str_recipe_target())
-
-        self.makefile += self.str_make_clean()
-        return self.makefile
-
-    def handle_variable(self, settings):  # variable
+    def handle_variable(self, settings):
+        '''Dealing with make variables'''
         var_string = ""
         if self.header:
-            string = " ".join(self.settings.get("include_dir"))
+            string = " ".join(self.header)
             header = "$(addprefix -I,$(HDR_DIR))"
             self.variables["HDR_DIR"] = string
             self.variables["HEADER"] = header
 
-            var_string += "HDR_DIR = {}\nHEADER = {}\n".format(string, header)
+            var_string += "HDR_DIR = %s\nHEADER = %s\n" % (string, header)
 
         if settings.get("lib_dir") and settings.get("lib_names"):
             lib_dir = " ".join(settings.get("lib_dir"))
@@ -66,78 +63,67 @@ class Makerfile():
             self.variables["LIB_NAMES"] = libnames
             self.variables["Library"] = libn
 
-            var_string += "LIB_DIR = {}\nLIB = {}\n".format(lib_dir, lib)
-            var_string += "LIB_NAMES = {}\nLIBRARY = {}\n".format(
-                libnames, libn)
+            var_string += "LIB_DIR = %s\nLIB = %s\n" % (lib_dir, lib)
+            var_string += "LIB_NAMES = %s\nLIBRARY = %s\n" % (libnames, libn)
 
         if settings.get("additional_flags"):
             self.variables["CCOPTION"] = " ".join(
                 settings.get("additional_flags"))
             self.variables["FLAGS"] = "$(addprefix -,$(FLAGS))"
 
-            var_string += "CCOPTION = {}\n".format(
-                self.variables.get("CCOPTION"))
+            var_string += "CCOPTION = %s\n" % self.variables.get("CCOPTION")
             var_string += "FLAGS = $(addprefix -,$(CCOPTION))\n"
 
         if settings.get("cc"):
             self.variables["CC"] = settings.get("cc")
-            var_string += "CC = {}\n".format(settings.get("cc"))
+            var_string += "CC = %s\n" % settings.get("cc")
         else:
-            var_string += "CC = {}\n".format("g++")
+            var_string += "CC = %s\n" % "g++"
 
-        if settings.get("projec_name"):
-            if sys.platform == "win32":
-                self.variables["main_file"] = settings.get(
-                    "main_file") + ".exe"
-            else:
-                self.variables["main_file"] = settings.get(
-                    "main_file") + ".out"
+        exe_ext = '.exe' if sublime.platform == 'windows' else '.out'
+        if settings.get("project_name"):
+            self.variables["main_file"] = settings.get(
+                "project_name") + exe_ext
 
         elif settings.get("main_file"):
-            if sys.platform == "win32":
-                self.variables["main_file"] = settings.get(
-                    "main_file") + ".exe"
-            else:
-                self.variables["main_file"] = settings.get(
-                    "main_file") + ".out"
+            self.variables["main_file"] = settings.get(
+                "main_file") + exe_ext
         else:
-            if sys.platform == "win32":
-                self.variables["main_file"] = "output.exe"
-            else:
-                self.variables["main_file"] = "output.out"
+            self.variables["main_file"] = "output" + exe_ext
 
         if self.obj:
-            var_string += "OBJ_DIR = {}\n".format(self.obj)
+            var_string += "OBJ_DIR = %s\n" % self.obj
 
         if self.sources:
             self.variables["OBJ"] = ""
             for x in self.sources:
                 self.variables["OBJ"] += x.replace(".cpp", ".o") + " "
 
-            var_string += "OBJ = {}\n".format(self.variables.get("OBJ"))
+            var_string += "OBJ = %s\n" % self.variables.get("OBJ")
 
         if self.build:
             self.variables["BUILD_DIR"] = self.build
-            var_string += "BUILD_DIR = {}\n".format(self.build)
+            var_string += "BUILD_DIR = %s\n" % self.build
 
         if self.src:
             self.variables["SRC_DIR"] = self.src
-            var_string += "SRC_DIR = {}\n".format(self.src)
+            var_string += "SRC_DIR = %s\n" % self.src
 
         return var_string
 
     def str_main_file(self):
+        '''Generate final output recipe for make'''
         string = ""
         if self.build:
-            string += "{2}" + self.path_sep # build dir
+            string += "{2}" + os.sep  # build dir
 
-        string += "{0}: {1} \n\t $(CC) "
+        string += "{0}: {1} \n\t$(CC) "
 
         if self.variables.get("FLAGS"):
             string += "$(FLAGS) "
 
         if self.build:
-            string += "{1} -o {2}" + self.path_sep + "{0} "
+            string += "{1} -o {2}" + os.sep + "{0} "
         else:
             string += "{1} -o {0} "
 
@@ -145,7 +131,7 @@ class Makerfile():
             string += "$(LIB) $(LIBRARY)"
 
         if self.obj:
-            objs = "$(addprefix $(OBJ_DIR)\,$(OBJ))"
+            objs = "$(addprefix $(OBJ_DIR){},$(OBJ))".format(os.sep)
         else:
             objs = "$(OBJ)"
 
@@ -153,78 +139,57 @@ class Makerfile():
             self.variables.get("main_file"), objs, "$(BUILD_DIR)")
         return main_exe
 
+    def str_make_run(self):
+        return 'run: {0}{1}{2}\n\t{0}{1}{2}\n'.format(
+            '$(BUILD_DIR)', os.sep, self.variables.get("main_file"))
+
     def str_make_clean(self):
-        temp = "clean: \n\t"
-
-        del_command = "rm "
-
-        if sublime.platform() == 'windows':
-            del_command = "del /Q "
+        '''Generate `clean` recipe for make'''
+        clean_content = ['clean: ']
+        del_command = "rm " if sublime.platform() != 'windows' else 'del /Q'
 
         if bool(self.settings.get("clean")):
             for i in self.settings.get("clean"):
-                if sublime.platform() != 'windows':
-                    i = i.replace('\\', '/')
-                temp += del_command + i + "\n\t"
+                i = i.replace('\\', os.sep)
+                clean_content.append(del_command + i)
         else:
-            if sublime.platform() != 'windows':
-                temp += del_command + "$(OBJ_DIR)/*.o\n\t"
-                temp += del_command + "$(BUILD_DIR)/*.out\n"
-            else:
-                temp += del_command + "$(OBJ_DIR)\\*.o\n\t"
-                temp += del_command + "$(BUILD_DIR)\\*.exe\n"
+            clean_content.append(del_command + "$(OBJ_DIR)%s*.o" % os.sep)
+            clean_content.append(del_command + "$(BUILD_DIR)%s*.out" % os.sep)
 
-        return temp
-
-    def get_source_files(self, src_dir):
-        back_out = False
-        if src_dir:
-            try:
-                os.chdir(src_dir)
-                back_out = True
-            except FileNotFoundError:
-                print("Couldn't find the Folder: ", src_dir)
-
-        file_sources = glob.glob("*.cpp")
-
-        if back_out:
-            os.chdir("..")
-        return file_sources
+        return "\n\t".join(clean_content)  # strip last `\t`
 
     def str_recipe_target(self):
-        tempmk = ""
+        tempmk = []
         if self.sources:
             for x in self.sources:
-                tempmk += self.string_template.format(
+                tempmk.append(self.string_template.format(
                     x.replace(".cpp", ".o"), x, "$(OBJ_DIR)", "$(SRC_DIR)")
-                tempmk += "\n\n"
-            return tempmk
+                )
+            return '\n\n'.join(tempmk) + '\n\n'
         else:
             print("No Source files found")
             return None
 
-    # generate a template string based on the variables passed for the object
     def make_string_template(self):
+        '''Generate template string based on variables passed for the object'''
         tmp = ""
         if self.obj:
-            tmp += "{2}" + self.path_sep  # obj dir
+            tmp += "{2}" + os.sep  # obj dir
 
         tmp += "{0}: "  # object .o
 
         if self.src:
-            tmp += "{3}" + self.path_sep  # src dir
+            tmp += "{3}" + os.sep  # src dir
 
-        tmp += "{1} \n\t $(CC) $(FLAGS) -c "  # recipe for object
-
-        # print(self.src)
+        tmp += "{1}\n\t$(CC) $(FLAGS) -c "  # recipe for object
 
         if self.src:
-            tmp += "{3}" + self.path_sep 
+            tmp += "{3}" + os.sep
 
         tmp += "{1} "
 
         if self.obj:
-            tmp += "-o {2}" + self.path_sep + "{0} "
+            tmp += "-o {2}" + os.sep + "{0} "
         else:
             tmp += "-o {0} "
 
@@ -232,6 +197,22 @@ class Makerfile():
             tmp += "$(HEADER)"
 
         self.string_template = tmp
+
+    def get_source_files(self, src_dir):
+        '''Find all .cpp source file in src_dir'''
+        back_out = './'
+        if src_dir:
+            try:
+                back_out = os.getcwd()
+                os.chdir(src_dir)
+            except FileNotFoundError:
+                sublime.error_message("Couldn't find source Folder: " + src_dir)
+
+        file_sources = glob.glob("*.cpp")
+
+        os.chdir(back_out)
+
+        return file_sources
 
     # setter methods
     def set_src(self, src_dir):
